final/requirements.txt
absl-py==1.3.0
actionlib==1.13.2
angles==1.9.13
asttokens==2.2.1
backcall==0.2.0
base-local-planner==1.17.2
bondpy==1.8.6
cachetools==5.2.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
executing==1.2.0
fonttools==4.38.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
image-geometry==1.16.2
importlib-metadata==6.0.0
interactive-markers==1.12.0
ipython==8.8.0
jedi==0.18.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.4
laser-geometry==1.6.7
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
message-filters==1.15.15
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.3.0.36
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-qt-binding==0.4.4
pytz==2022.7
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.28.1
requests-oauthlib==1.3.1
resource-retriever==1.12.7
rosbag==1.15.15
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.15
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospy==1.15.15
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rtabmap-ros==0.20.20
rviz==1.14.19
scipy==1.10.0
seaborn==0.12.2
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
thop==0.1.1.post2209072238
topic-tools==1.15.15
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
xacro==1.14.14
zipp==3.11.0
